Move TTestTokenExchangeServer to reusable library
