+++ ydb/public/sdk/cpp/tests/unit/client/oauth2_token_exchange/credentials_ut.cpp
@@ -1,13 +1,9 @@
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/types/credentials/oauth2_token_exchange/credentials.h>
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/types/credentials/oauth2_token_exchange/from_file.h>
-#include "jwt_check_helper.h"
+#include <ydb/public/sdk/cpp/tests/unit/client/oauth2_token_exchange/helpers/test_token_exchange_server.h>
 
 #include <library/cpp/string_utils/base64/base64.h>
 
-#include <library/cpp/cgiparam/cgiparam.h>
-#include <library/cpp/http/misc/parsed_request.h>
-#include <library/cpp/http/server/http.h>
-#include <library/cpp/http/server/response.h>
 #include <library/cpp/json/json_writer.h>
 #include <library/cpp/testing/unittest/registar.h>
 #include <library/cpp/testing/unittest/tests_data.h>
@@ -23,164 +19,6 @@ extern const std::string TestRSAPublicKeyContent;
 extern const std::string TestECPrivateKeyContent;
 extern const std::string TestECPublicKeyContent;
 extern const std::string TestHMACSecretKeyBase64Content;
-
-class TTestTokenExchangeServer: public THttpServer::ICallBack {
-public:
-    struct TCheck {
-        bool ExpectRequest = true;
-        HttpCodes StatusCode = HTTP_OK;
-        TCgiParameters ExpectedInputParams;
-        std::optional<TCgiParameters> InputParams;
-        std::string Response;
-        std::string ExpectedErrorPart;
-        std::string Error;
-        std::optional<TJwtCheck> SubjectJwtCheck;
-        std::optional<TJwtCheck> ActorJwtCheck;
-
-        void Check() {
-            UNIT_ASSERT_C(InputParams || !ExpectRequest, "Request error: " << Error);
-            if (InputParams) {
-                if (SubjectJwtCheck || ActorJwtCheck) {
-                    TCgiParameters inputParamsCopy = *InputParams;
-                    if (SubjectJwtCheck) {
-                        std::string subjectJwt;
-                        UNIT_ASSERT(inputParamsCopy.Has("subject_token"));
-                        UNIT_ASSERT(inputParamsCopy.Has("subject_token_type", "urn:ietf:params:oauth:token-type:jwt"));
-                        subjectJwt = inputParamsCopy.Get("subject_token");
-                        inputParamsCopy.Erase("subject_token");
-                        inputParamsCopy.Erase("subject_token_type");
-                        SubjectJwtCheck->Check(subjectJwt);
-                    }
-                    if (ActorJwtCheck) {
-                        std::string actorJwt;
-                        UNIT_ASSERT(inputParamsCopy.Has("actor_token"));
-                        UNIT_ASSERT(inputParamsCopy.Has("actor_token_type", "urn:ietf:params:oauth:token-type:jwt"));
-                        actorJwt = inputParamsCopy.Get("actor_token");
-                        inputParamsCopy.Erase("actor_token");
-                        inputParamsCopy.Erase("actor_token_type");
-                        ActorJwtCheck->Check(actorJwt);
-                    }
-                    UNIT_ASSERT_VALUES_EQUAL(ExpectedInputParams.Print(), inputParamsCopy.Print());
-                } else {
-                    UNIT_ASSERT_VALUES_EQUAL(ExpectedInputParams.Print(), InputParams->Print());
-                }
-            }
-
-            if (!ExpectedErrorPart.empty()) {
-                UNIT_ASSERT_STRING_CONTAINS(Error, ExpectedErrorPart);
-            } else {
-                UNIT_ASSERT(Error.empty());
-            }
-        }
-
-        void Reset() {
-            Error.clear();
-            InputParams = std::nullopt;
-        }
-    };
-
-    class TRequest: public TRequestReplier {
-    public:
-        explicit TRequest(TTestTokenExchangeServer* server)
-            : Server(server)
-        {
-        }
-
-        bool DoReply(const TReplyParams& params) override {
-            with_lock (Server->Lock) {
-                const TParsedHttpFull parsed(params.Input.FirstLine());
-                UNIT_ASSERT_VALUES_EQUAL(parsed.Path, "/exchange/token");
-                const std::string bodyStr = params.Input.ReadAll();
-
-                Server->Check.InputParams.emplace(bodyStr);
-                THttpResponse resp(Server->Check.StatusCode);
-                resp.SetContent(TString{Server->Check.Response});
-                resp.OutTo(params.Output);
-                return true;
-            }
-        }
-
-    public:
-        TTestTokenExchangeServer* Server = nullptr;
-    };
-
-    TTestTokenExchangeServer()
-        : HttpOptions(PortManager.GetPort())
-        , HttpServer(this, HttpOptions)
-    {
-        HttpServer.Start();
-    }
-
-    TClientRequest* CreateClient() override {
-        return new TRequest(this);
-    }
-
-    std::string GetEndpoint() const {
-        return TStringBuilder() << "http://localhost:" << HttpOptions.Port << "/exchange/token";
-    }
-
-    void Run(const std::function<void()>& f, bool checkExpectations = true) {
-        Check.Reset();
-        try {
-            f();
-        } catch (const std::exception& ex) {
-            Check.Error = ex.what();
-        }
-
-        if (checkExpectations) {
-            CheckExpectations();
-        }
-    }
-
-    void Run(const TOauth2TokenExchangeParams& params, const std::string& expectedToken = {}, bool checkExpectations = true) {
-        std::string token;
-        Run([&]() {
-            auto factory = CreateOauth2TokenExchangeCredentialsProviderFactory(params);
-            if (!expectedToken.empty()) {
-                token = factory->CreateProvider()->GetAuthInfo();
-            }
-        },
-        checkExpectations);
-
-        if (!expectedToken.empty()) {
-            UNIT_ASSERT_VALUES_EQUAL(expectedToken, token);
-        }
-    }
-
-    void RunFromConfig(const std::string& fileName, const std::string& expectedToken = {}, bool checkExpectations = true, const std::string& explicitTokenEndpoint = {}) {
-        std::string token;
-        Run([&]() {
-            auto factory = CreateOauth2TokenExchangeFileCredentialsProviderFactory(fileName, explicitTokenEndpoint);
-            if (!expectedToken.empty()) {
-                token = factory->CreateProvider()->GetAuthInfo();
-            }
-        },
-        checkExpectations);
-
-        if (!expectedToken.empty()) {
-            UNIT_ASSERT_VALUES_EQUAL(expectedToken, token);
-        }
-    }
-
-    void CheckExpectations() {
-        with_lock (Lock) {
-            Check.Check();
-        }
-    }
-
-    void WithLock(const std::function<void()>& f) {
-        with_lock (Lock) {
-            f();
-        }
-    }
-
-public:
-    TAdaptiveLock Lock;
-    TPortManager PortManager;
-    THttpServer::TOptions HttpOptions;
-    THttpServer HttpServer;
-    TCheck Check;
-};
 
 template <class TParent>
  struct TJsonFillerArray {
