+++ ydb/public/sdk/cpp/tests/unit/client/oauth2_token_exchange/helpers/ya.make
@@ -1,17 +1,17 @@
-UNITTEST()
+LIBRARY()
 
 SRCS(
-    credentials_ut.cpp
-    jwt_token_source_ut.cpp
+    test_token_exchange_server.cpp
 )
 
 PEERDIR(
     contrib/libs/jwt-cpp
+    library/cpp/cgiparam
+    library/cpp/http/misc
     library/cpp/http/server
     library/cpp/json
     library/cpp/string_utils/base64
     ydb/public/sdk/cpp/src/client/types/credentials/oauth2_token_exchange
-    ydb/public/sdk/cpp/tests/unit/client/oauth2_token_exchange/helpers
 )
 
 END()
