+++ ydb/public/sdk/cpp/tests/unit/client/oauth2_token_exchange/jwt_token_source_ut.cpp
@@ -1,5 +1,5 @@
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/types/credentials/oauth2_token_exchange/jwt_token_source.h>
-#include "jwt_check_helper.h"
+#include <ydb/public/sdk/cpp/tests/unit/client/oauth2_token_exchange/helpers/jwt_check_helper.h>
 
 #include <library/cpp/testing/unittest/registar.h>
